Move docling related methods to docling_utils.py
