+++ ai-services/assets/catalog/podman/values.yaml
@@ -4,7 +4,7 @@ ui:
 
 backend:
   port: ""
-  image: icr.io/ai-services-cicd/ai-services:v0.0.112
+  image: icr.io/ai-services-cicd/ai-services:v0.0.113
   runtime: ""
   adminPasswordHash: ""
   podman:
